Fixed vulnerability with symlink/junction files
Potential for attack if any of the XML files that ACAT writes to is a symlink, or if the parent folder is a junction. When ACAT writes to the file, it actually overwrites the file that is the target of the symlink.  This vulnerability can be exploited by bad actors.  With this fix, ACAT will exit the app if it finds that the files being written to are synlinks or if the parent folder is a junction